+++ src/main/java/nz/co/mirality/colony4cc/peripheral/PocketColonyPeripheral.java
@@ -3,7 +3,6 @@ package nz.co.mirality.colony4cc.peripheral;
 import dan200.computercraft.api.peripheral.IPeripheral;
 import net.minecraft.entity.Entity;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.vector.Vector3d;
 import net.minecraft.world.World;
 
 import javax.annotation.Nullable;